moving methods to correct places
