+++ src/main/java/harmony/mastermind/logic/commands/RelocateCommand.java
@@ -17,9 +17,10 @@ import harmony.mastermind.model.tag.Tag;
 import harmony.mastermind.model.tag.UniqueTagList;
 import harmony.mastermind.model.task.*;
 import harmony.mastermind.storage.StorageManager;
+import harmony.mastermind.storage.*;
 
 /**
- * @author A0139194X
+ * @@author A0139194X
  * Relocates save location
  */
 public class RelocateCommand extends Command{
@@ -34,7 +35,6 @@ public class RelocateCommand extends Command{
     public static final String MESSAGE_SUCCESS = "Relocated save location to %1$s";
     public static final String MESSAGE_INVALID_INPUT = "%1$s is not valid.";
     public static final String MESSAGE_UNWRITTABLE_FOLDER = "%1$s is not writtable.";
-
 
     private final String newFilePath;
 
@@ -51,8 +51,8 @@ public class RelocateCommand extends Command{
     public CommandResult execute() {
         assert model != null;
         try {
-            checkSaveLocation(newFilePath);
-            checkWrittableDirectory(newFilePath);
+            storageManager.checkSaveLocation(newFilePath);
+            storageManager.checkWrittableDirectory(newFilePath);
             model.relocateSaveLocation(newFilePath);
             return new CommandResult(COMMAND_WORD, String.format(MESSAGE_SUCCESS, newFilePath));
         } catch (FolderDoesNotExistException fdnee) {
@@ -61,21 +61,4 @@ public class RelocateCommand extends Command{
             return new CommandResult(COMMAND_WORD, String.format(MESSAGE_UNWRITTABLE_FOLDER, newFilePath));
         }
     }
-    
-    //@@author A0139194X
-    public void checkSaveLocation(String newFilePath) throws FolderDoesNotExistException {
-        Path filePath = Paths.get(newFilePath);
-        if (!Files.exists(filePath)) {
-            throw new FolderDoesNotExistException(newFilePath + " does not exist");
-        }
-    }
-    
-    //@@author A0139194X
-    public void checkWrittableDirectory(String newFilePath) throws UnwrittableFolderException {
-        File newFile = new File(newFilePath);
-        if (!(newFile.isDirectory() && newFile.canWrite())) {
-            throw new UnwrittableFolderException(newFilePath + " is not writtable.");
-        }
-    }
-    
 }
